cc-4485 revert changes

+++ Bundles/ProductMeasurementUnit/src/Spryker/Zed/ProductMeasurementUnit/Business/ProductMeasurementUnitFacade.php
@@ -45,9 +45,9 @@ class ProductMeasurementUnitFacade extends AbstractFacade implements ProductMeas
      *
      * @param \Generated\Shared\Transfer\ItemTransfer $itemTransfer
      *
-     * @return float
+     * @return int
      */
-    public function calculateQuantityNormalizedSalesUnitValue(ItemTransfer $itemTransfer): float
+    public function calculateQuantityNormalizedSalesUnitValue(ItemTransfer $itemTransfer): int
     {
         return $this->getFactory()
             ->createProductMeasurementSalesUnitValue()
